Conformance Update
Added:
 - Mod Spec as a normative document
 - Terms and definitions related to conformance
 - Added a discussion of conformance to clause 2
Also: updated text of requirements classes to have dependencies on requirements classes, not conformance classes.

Finally, I find that the Mod Spec is not clear on the difference between a conformance class and a requirements class. In my discussion I assert that conformance classes describe capabilities (or functionality) while the requirments classes define how to validate that capablity. I beleve this is in-line with the intent of the Mod Spec if not the wording.